requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
Jinja2==3.1.2
openpyxl==3.1.2
reportlab==4.0.7
gunicorn==21.2.0
python-dateutil==2.8.2

# CSRF protection (planned)
Flask-WTF==1.2.1

# File locking for concurrent access to JSON DB
portalocker==2.8.2

# PostgreSQL Support
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
